// src/models/menu-item.ts
export class MenuItem {
    producto: string;
    categoria: string;
    alergenos: string[];
    descripcion: string;
    foto_tapa: string;
    foto_tostada: string;
    foto_media_tabla: string;
    foto_tabla: string;
    tapa: number;
    tostada: number;
    media_tabla: number;
    tabla: number;
}